Improved TagsController#match for tag tables; enhanced query support in token_input_tag helper; modified Tag#absorb to support making synonyms

+++ app/models/tag.rb
@@ -178,7 +178,9 @@ class Tag < ActiveRecord::Base
     end
   end
 
-  def absorb other
+  # Use this tag instead of 'other', i.e., absorb its taggings, referents, etc.
+  # Either delete the other, or make it a synonym, according to 'keep'
+  def absorb other, delete=true
     return other if other.id == id
     # Normal procedure:
     TaggingServices.change_tag(other.id, self.id)
@@ -198,12 +200,25 @@ class Tag < ActiveRecord::Base
       # Failure: copy errors into the original record and return it
       self.errors.each { |k, v| other.errors[k] = v }
       other
-    else
+    elsif delete
       Tag.transaction do
         other.destroy
         self.save
       end
       self
+    else
+      # The other may clash with an existing tag of the target type,
+      # So get either the other, or its equivalent
+      other = other.project tagtype
+      if other != self
+        # No need to monkey with referents if the projection devolved on us
+        Referent.express self unless primary_meaning || referent_ids.present?
+        # Leave the other as a synonym
+        other.referent_ids = other.referent_ids + self.referent_ids
+        other.primary_meaning ||= self.primary_meaning
+        other.save
+      end
+      other
     end
   end
 
@@ -225,8 +240,8 @@ class Tag < ActiveRecord::Base
 
   # Return the tag's name with a marker of its type, to clear up ambiguities
   def typedname include_type=false, include_ref=false
-    return name unless include_type && (typenum > 0)
-    referent_str = (include_ref && referent_id && (' '+referent_id.to_s)) || ''
+    return name unless include_type
+    referent_str = "#{primary_meaning.model_name.human}##{primary_meaning.id}" if include_ref && referent_id
     %Q{#{name} [#{typename}#{referent_str}]}
   end
